changed expected duration type for attack scripts

+++ new_scripts/Attacker/empty_connection_dos.py
@@ -30,7 +30,7 @@ def main(username, password, duration=10):
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="launch an empty_con_dos attack from the local machine")
-    parser.add_argument("-d", "--duration", type=int, help="duration in seconds")
+    parser.add_argument("-d", "--duration", type=float, help="duration in seconds")
     parser.add_argument("-u", "--username", type=str, help="username to connect as")
     parser.add_argument("-p", "--password", type=str, help="password to connect with")
     args = parser.parse_args()
